refactor(workflow): extract contract loading
