+++ src/workflow.ts
@@ -5,7 +5,14 @@ import { parse } from "yaml";
 import { BUILTIN_WORKFLOW_TEMPLATES } from "./builtin-templates.js";
 import type { WorkflowFormat } from "./config-schemas.js";
 import { isPathInside } from "./path-safety.js";
-import { validatePromptTemplateExpressions } from "./workflow/autonomous-prompt.js";
+import {
+  parseWorkflowContract,
+  projectWorkflowReferences,
+  resolveWorkflowFormat,
+  selectProjectWorkflow,
+  validateWorkflowTemplate
+} from "./workflow/contract-loading.js";
+import type { WorkflowContract } from "./workflow/contract-loading.js";
 import type {
   ExpandedWorkflow,
   ExpandedWorkflowState,
@@ -32,6 +39,13 @@ export type {
   RenderedAutonomousPrompt,
   RunEvidencePaths
 } from "./workflow/autonomous-prompt.js";
+export {
+  loadWorkflowContract,
+  parseWorkflowContract,
+  validateWorkflowContract,
+  validateWorkflowTemplate
+} from "./workflow/contract-loading.js";
+export type { WorkflowContract } from "./workflow/contract-loading.js";
 export type {
   ExpandedWorkflow,
   ExpandedWorkflowState,
@@ -42,13 +56,6 @@ export type {
   WorkflowSourceKind,
   WorkflowTransition
 } from "./workflow/types.js";
-
-export type WorkflowContract = {
-  body: string;
-  contentHash: string;
-  errors: string[];
-  path: string;
-};
 
 type WorkflowTemplateInputType =
   | "boolean"
@@ -103,18 +110,6 @@ export type ProjectWorkflowLoadResult = {
   workflowPath: string | null;
 };
 
-const serviceDiscoveryFrontMatterKeys = new Set([
-  "agent",
-  "issue_filters",
-  "priority",
-  "projects",
-  "provider",
-  "providers",
-  "tracker",
-  "workflow",
-  "workspace"
-]);
-
 const actionKinds = new Set<WorkflowActionKind>([
   "agent",
   "close_issue",
@@ -145,13 +140,6 @@ const completionPredicateKeys = new Set([
 const terminalStates = new Set(["blocked", "failure", "success"]);
 
 const tagPattern = /{{\s*([^{}]+?)\s*}}/g;
-
-export async function loadWorkflowContract(
-  workflowPath: string
-): Promise<WorkflowContract> {
-  const contents = await readFile(workflowPath, "utf8");
-  return parseWorkflowContract(contents, workflowPath);
-}
 
 export async function loadExpandedWorkflow(
   workflowPath: string,
@@ -298,84 +286,6 @@ export function explainWorkflow(workflow: ExpandedWorkflow): string {
   return `${lines.join("\n")}\n`;
 }
 
-export async function validateWorkflowContract(
-  workflowPath: string
-): Promise<string[]> {
-  let contents: string;
-
-  try {
-    contents = await readFile(workflowPath, "utf8");
-  } catch (error) {
-    return [`workflow contract not found at ${workflowPath}: ${errorMessage(error)}`];
-  }
-
-  const workflow = parseWorkflowContract(contents, workflowPath);
-  const errors = [...workflow.errors];
-
-  if (workflow.body.trim().length === 0) {
-    errors.push(`workflow contract at ${workflowPath} must not be empty`);
-  }
-
-  errors.push(...validateWorkflowTemplate(workflow.body, workflowPath));
-  return errors;
-}
-
-export function parseWorkflowContract(
-  contents: string,
-  workflowPath: string
-): WorkflowContract {
-  const lines = contents.split(/\r?\n/);
-
-  if (lines[0]?.trim() !== "---") {
-    return {
-      body: contents,
-      contentHash: contentHash(contents),
-      errors: [],
-      path: workflowPath
-    };
-  }
-
-  const closingLine = lines.findIndex(
-    (line, index) => index > 0 && line.trim() === "---"
-  );
-  if (closingLine === -1) {
-    return {
-      body: "",
-      contentHash: contentHash(contents),
-      errors: [`workflow front matter at ${workflowPath} is missing a closing ---`],
-      path: workflowPath
-    };
-  }
-
-  const frontMatterSource = lines.slice(1, closingLine).join("\n");
-  const errors: string[] = [];
-  const frontMatter = parseFrontMatter(frontMatterSource, workflowPath, errors);
-
-  if (frontMatter !== undefined) {
-    for (const key of Object.keys(frontMatter)) {
-      if (serviceDiscoveryFrontMatterKeys.has(key)) {
-        errors.push(
-          `workflow front matter at ${workflowPath} must not define service config key ${key}`
-        );
-      }
-    }
-  }
-
-  return {
-    body: lines.slice(closingLine + 1).join("\n"),
-    contentHash: contentHash(contents),
-    errors,
-    path: workflowPath
-  };
-}
-
-export function validateWorkflowTemplate(
-  template: string,
-  workflowPath: string
-): string[] {
-  return validatePromptTemplateExpressions(template, workflowPath);
-}
-
 export async function expandWorkflowDefinition(
   contents: string,
   workflowPath: string,
@@ -419,33 +329,6 @@ export async function expandWorkflowDefinition(
   };
 }
 
-type ResolvedWorkflowFormat =
-  | { kind: "markdown" | "raw_fsm" }
-  | { error: string; kind: "error" };
-
-function resolveWorkflowFormat(
-  format: WorkflowFormat,
-  workflowPath: string
-): ResolvedWorkflowFormat {
-  if (format === "markdown") {
-    return { kind: "markdown" };
-  }
-  if (format === "raw_fsm") {
-    return { kind: "raw_fsm" };
-  }
-  const extension = path.extname(workflowPath).toLowerCase();
-  if (extension === ".md") {
-    return { kind: "markdown" };
-  }
-  if (extension === ".yaml" || extension === ".yml" || extension === ".json") {
-    return { kind: "raw_fsm" };
-  }
-  return {
-    error: `workflow at ${workflowPath} has no recognized extension (.md, .yaml, .yml, .json); declare format explicitly`,
-    kind: "error"
-  };
-}
-
 function emptyExpandedWorkflow(
   contents: string,
   workflowPath: string,
@@ -459,122 +342,6 @@ function emptyExpandedWorkflow(
     states: [],
     templateFiles: []
   };
-}
-
-function projectWorkflowReferences(
-  rawProjects: unknown[],
-  configPath: string,
-  errors: string[]
-): Array<{ name: string; workflowFormat: WorkflowFormat; workflowPath: string }> {
-  const projects: Array<{
-    name: string;
-    workflowFormat: WorkflowFormat;
-    workflowPath: string;
-  }> = [];
-  for (const [index, rawProject] of rawProjects.entries()) {
-    if (!isRecord(rawProject)) {
-      errors.push(`projects.${index} in ${configPath} must be a mapping`);
-      continue;
-    }
-    const name = stringProperty(rawProject, "name");
-    if (name === undefined) {
-      errors.push(`projects.${index}.name in ${configPath} must be a non-empty string`);
-      continue;
-    }
-    const reference = parseWorkflowReference(
-      rawProject.workflow,
-      `projects.${name}.workflow`,
-      configPath,
-      errors
-    );
-    if (reference === undefined) {
-      continue;
-    }
-    projects.push({
-      name,
-      workflowFormat: reference.format,
-      workflowPath: reference.path
-    });
-  }
-  return projects;
-}
-
-function parseWorkflowReference(
-  rawWorkflow: unknown,
-  fieldLabel: string,
-  configPath: string,
-  errors: string[]
-): { format: WorkflowFormat; path: string } | undefined {
-  if (typeof rawWorkflow === "string") {
-    const trimmed = rawWorkflow.trim();
-    if (trimmed.length === 0) {
-      errors.push(`${fieldLabel} in ${configPath} must be a non-empty path`);
-      return undefined;
-    }
-    return { format: "auto", path: trimmed };
-  }
-  if (isRecord(rawWorkflow)) {
-    const pathValue = stringProperty(rawWorkflow, "path");
-    if (pathValue === undefined) {
-      errors.push(`${fieldLabel}.path in ${configPath} must be a non-empty path`);
-      return undefined;
-    }
-    const formatRaw = rawWorkflow.format;
-    let format: WorkflowFormat = "auto";
-    if (formatRaw !== undefined) {
-      if (
-        formatRaw === "markdown" ||
-        formatRaw === "raw_fsm" ||
-        formatRaw === "auto"
-      ) {
-        format = formatRaw;
-      } else {
-        errors.push(
-          `${fieldLabel}.format in ${configPath} must be one of markdown, raw_fsm, auto`
-        );
-        return undefined;
-      }
-    }
-    return { format, path: pathValue };
-  }
-  errors.push(`${fieldLabel} in ${configPath} must be a non-empty path or mapping`);
-  return undefined;
-}
-
-function selectProjectWorkflow(
-  projects: Array<{
-    name: string;
-    workflowFormat: WorkflowFormat;
-    workflowPath: string;
-  }>,
-  requestedProject: string | undefined,
-  configPath: string,
-  errors: string[]
-):
-  | { name: string; workflowFormat: WorkflowFormat; workflowPath: string }
-  | undefined {
-  if (requestedProject !== undefined) {
-    const selected = projects.find((project) => project.name === requestedProject);
-    if (selected === undefined) {
-      errors.push(
-        `project ${requestedProject} is not defined in service config ${configPath}`
-      );
-    }
-    return selected;
-  }
-
-  if (projects.length === 1) {
-    return projects[0];
-  }
-
-  if (projects.length === 0) {
-    errors.push(`service config ${configPath} does not contain a project workflow`);
-  } else {
-    errors.push(
-      `service config ${configPath} has ${projects.length} projects; pass --project`
-    );
-  }
-  return undefined;
 }
 
 function parseExplicitWorkflowDefinition(
@@ -1595,26 +1362,6 @@ function uniqueStrings(values: string[]): string[] {
 
 function isPathSafeIdentifier(value: string): boolean {
   return /^[A-Za-z_][A-Za-z0-9_.-]*$/.test(value);
-}
-
-function parseFrontMatter(
-  source: string,
-  workflowPath: string,
-  errors: string[]
-): Record<string, unknown> | undefined {
-  try {
-    const parsed: unknown = parse(source) ?? {};
-    if (isRecord(parsed)) {
-      return parsed;
-    }
-    errors.push(`workflow front matter at ${workflowPath} must be a mapping`);
-    return undefined;
-  } catch (error) {
-    errors.push(
-      `workflow front matter at ${workflowPath} could not be parsed: ${errorMessage(error)}`
-    );
-    return undefined;
-  }
 }
 
 function contentHash(contents: string): string {
