+++ tests/workflow.test.ts
@@ -8,14 +8,11 @@ import { decideNextStep } from "../src/lifecycle/state-machine-dispatch.js";
 import {
   explainWorkflow,
   loadExpandedWorkflow,
-  loadWorkflowContract,
-  renderAutonomousPrompt,
   validateExpandedWorkflowReferences
 } from "../src/workflow.js";
 import type { ExpandedWorkflow } from "../src/workflow.js";
 
 const tempRoots: string[] = [];
-const DEFAULT_CODEX_COMMAND = `codex -p symphonika -c sandbox_mode=danger-full-access -c approval_policy=never --dangerously-bypass-approvals-and-sandbox app-server`;
 
 async function makeTempRoot(): Promise<string> {
   const root = await mkdtemp(path.join(tmpdir(), "symphonika-workflow-test-"));
@@ -29,62 +26,6 @@ afterEach(async () => {
       rm(root, { force: true, recursive: true })
     )
   );
-});
-
-describe("workflow contract loading", () => {
-  it("loads WORKFLOW.md body without front matter and carries its content hash into the rendered prompt", async () => {
-    const root = await makeTempRoot();
-    const workflowPath = path.join(root, "WORKFLOW.md");
-    await writeFile(
-      workflowPath,
-      [
-        "---",
-        "autonomy:",
-        "  max_turns: 8",
-        "---",
-        "Work on {{issue.title}}.",
-        ""
-      ].join("\n")
-    );
-
-    const workflow = await loadWorkflowContract(workflowPath);
-    const rendered = renderAutonomousPrompt({
-      branch: {
-        name: "sym/symphonika/7-render-prompts",
-        ref: "refs/heads/sym/symphonika/7-render-prompts"
-      },
-      issue: issueSnapshot(),
-      project: {
-        name: "symphonika"
-      },
-      provider: {
-        command: DEFAULT_CODEX_COMMAND,
-        name: "codex"
-      },
-      run: {
-        attempt: 1,
-        continuation: false,
-        id: "run-7"
-      },
-      template: workflow.body,
-      workflowContentHash: workflow.contentHash,
-      workflowPath: workflow.path,
-      workspace: {
-        path: "/state/workspaces/symphonika/issues/7-render-prompts",
-        previous_attempt: false,
-        root: "/state/workspaces/symphonika"
-      }
-    });
-
-    expect(workflow.errors).toEqual([]);
-    expect(workflow.body).toBe("Work on {{issue.title}}.\n");
-    expect(workflow.contentHash).toMatch(/^sha256:[a-f0-9]{64}$/);
-    expect(rendered.prompt).toContain(
-      "Work on Render autonomous prompts and persist run evidence."
-    );
-    expect(rendered.prompt).not.toContain("max_turns");
-    expect(rendered.workflowContentHash).toBe(workflow.contentHash);
-  });
 });
 
 describe("state machine workflow definitions", () => {
@@ -1782,131 +1723,6 @@ describe("built-in workflow templates", () => {
   });
 });
 
-describe("workflow format routing", () => {
-  it("auto-routes a .md file to the Markdown loader regardless of body content", async () => {
-    const root = await makeTempRoot();
-    const workflowPath = path.join(root, "WORKFLOW.md");
-    await writeFile(
-      workflowPath,
-      [
-        "workflow:",
-        "  name: looks_like_yaml",
-        "  initial: planning",
-        "  states:",
-        "    planning:",
-        "      action:",
-        "        kind: agent",
-        ""
-      ].join("\n")
-    );
-
-    const result = await loadExpandedWorkflow(workflowPath, "auto");
-
-    expect(result.errors).toEqual([]);
-    expect(result.workflow.source.kind).toBe("markdown");
-    expect(result.workflow.states.map((state) => state.id)).toEqual([
-      "run_agent",
-      "done"
-    ]);
-  });
-
-  it("auto-routes .yaml, .yml, and .json paths to the raw FSM loader", async () => {
-    const root = await makeTempRoot();
-    for (const extension of [".yaml", ".yml", ".json"]) {
-      const workflowPath = path.join(root, `workflow${extension}`);
-      const contents =
-        extension === ".json"
-          ? JSON.stringify({
-              workflow: {
-                initial: "done",
-                name: "json_workflow",
-                states: { done: { terminal: "success" } }
-              }
-            })
-          : [
-              "workflow:",
-              "  name: yaml_workflow",
-              "  initial: done",
-              "  states:",
-              "    done:",
-              "      terminal: success",
-              ""
-            ].join("\n");
-      await writeFile(workflowPath, contents);
-
-      const result = await loadExpandedWorkflow(workflowPath, "auto");
-      expect(result.errors).toEqual([]);
-      expect(result.workflow.source.kind).toBe("raw_fsm");
-      expect(result.workflow.states.map((state) => state.id)).toEqual(["done"]);
-    }
-  });
-
-  it("auto rejects an unknown extension and reports the supported set", async () => {
-    const root = await makeTempRoot();
-    const workflowPath = path.join(root, "workflow.txt");
-    await writeFile(workflowPath, "doesn't matter\n");
-
-    const result = await loadExpandedWorkflow(workflowPath, "auto");
-
-    expect(result.errors).toContain(
-      `workflow at ${workflowPath} has no recognized extension (.md, .yaml, .yml, .json); declare format explicitly`
-    );
-  });
-
-  it("explicit markdown format treats a YAML-shaped .yml file as Markdown content", async () => {
-    const root = await makeTempRoot();
-    const workflowPath = path.join(root, "workflow.yml");
-    await writeFile(
-      workflowPath,
-      [
-        "---",
-        "name: explicit_markdown",
-        "---",
-        "",
-        "Work on {{issue.title}}.",
-        ""
-      ].join("\n")
-    );
-
-    const result = await loadExpandedWorkflow(workflowPath, "markdown");
-
-    expect(result.errors).toEqual([]);
-    expect(result.workflow.source.kind).toBe("markdown");
-    expect(result.workflow.states.map((state) => state.id)).toEqual([
-      "run_agent",
-      "done"
-    ]);
-  });
-
-  it("explicit raw_fsm format errors when a .md file is not valid raw FSM YAML", async () => {
-    const root = await makeTempRoot();
-    const workflowPath = path.join(root, "WORKFLOW.md");
-    await writeFile(workflowPath, "# Just a markdown file\n");
-
-    const result = await loadExpandedWorkflow(workflowPath, "raw_fsm");
-
-    expect(result.errors).toContain(
-      `workflow definition at ${workflowPath} must define a top-level workflow mapping`
-    );
-  });
-
-  it("explicit raw_fsm format surfaces YAML parse errors regardless of extension", async () => {
-    const root = await makeTempRoot();
-    const workflowPath = path.join(root, "workflow.json");
-    await writeFile(workflowPath, "{ bogus yaml here: : :");
-
-    const result = await loadExpandedWorkflow(workflowPath, "raw_fsm");
-
-    expect(
-      result.errors.some((message) =>
-        message.startsWith(
-          `workflow definition at ${workflowPath} could not be parsed:`
-        )
-      )
-    ).toBe(true);
-  });
-});
-
 describe("validateExpandedWorkflowReferences", () => {
   it("returns no errors when every raw FSM agent prompt resolves to an existing file", async () => {
     const root = await makeTempRoot();
@@ -2055,18 +1871,3 @@ describe("validateExpandedWorkflowReferences", () => {
     expect(errors).toEqual([]);
   });
 });
-
-function issueSnapshot() {
-  return {
-    body: "Implement prompt evidence.",
-    created_at: "2026-04-28T10:00:00Z",
-    id: 700,
-    labels: ["agent-ready"],
-    number: 7,
-    priority: 99,
-    state: "open",
-    title: "Render autonomous prompts and persist run evidence",
-    updated_at: "2026-04-29T10:00:00Z",
-    url: "https://github.com/pmatos/symphonika/issues/7"
-  };
-}
